Created server and package for node js and moved code
Moved lib inside the pages subfolder so it can run with localhost.

Moved footer out into a separate html file which the others loaded in
(so we do not have to change the footer links on each page).

Changed footer links to navigate to actual facebook/twitter (linkedin
DNE to my knowledge).

Moved css into the styles file (noticed that a lot of the css was
starting to repeat and global css for these features would be nice).

Header html file was created, but not implemented since I could not
figure out how to switch the "active" class for each link. Also, I think
the header will remain fairly static.