Contact Form and Minor Style Update
Tweaked the header style for products page to display header on top.
Merged/removed infographic img. Reduced paragraph padding for
infographic to center it(magic number, will fix in refactoring).

Added a Contact Form. Currently, it does not work due to php and node
conflicts. We will need some external module or server for this.
However, the code is functional and based off of FormGet. The php sends
the request to justinreigel@yahoo.com (testing purposes only) and sends
a confirmation email to the user.